+++ www/modules/email/MessagesGrid.js
@@ -236,8 +236,6 @@ GO.email.MessagesGrid = function(config){
 						}
 					}
 				}
-
-
 			]
 		})
 	});
@@ -333,7 +331,6 @@ GO.email.MessagesGrid = function(config){
 			'->',
 			this.showUnreadButton,
 			this.searchField = new go.toolbar.SearchButton({
-				//store: config.store,
 				paramName: 'search',
 				hidden: config.hideSearch,
 				tools: [
@@ -381,7 +378,6 @@ GO.email.MessagesGrid = function(config){
 				],
 				listeners: {
 					search: function (me, v) {
-						debugger;
 						v = v.replace(/[\"\']/g, '');
 						config.store.baseParams['search'] = v;
 						config.store.load();